refactor: 동시성 테스트를 통합 테스트에 포함
